Добавлены файлы functions.cpp и header.h, переименован и изменен safr.cpp на main.cpp. Изменена функция 3 и добавлен default
