+++ project-razgrom/.vs/project-razgrom/v17/DocumentLayout.json
@@ -1,0 +1,12 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\\u0421\u0430\u0448\u043A\u0430\\project-razgrom\\project-razgrom\\",
+  "Documents": [],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": []
+    }
+  ]
+}
